bug fixes and additions

+++ lib/providers/prof_pic_provider.dart
@@ -9,41 +9,45 @@ import 'package:image_cropper/image_cropper.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:todo/utils/basic_utils.dart';
-import 'package:todo/utils/firebase_storage_utils.dart';
 
 class ProfPicProvider extends ChangeNotifier {
   String _profPicUrl = "";
   String curUserUid = BasicUtils().curUserUid;
   dynamic _profPic = const AssetImage("res/images/defaultProf.png");
   dynamic get profPic => _profPic;
+  bool _isUploading = false;
+  final storage = FirebaseStorage.instance;
+  bool get isUploading => _isUploading;
 
-  downloadProfPic() {
-    FirebaseFirestore.instance
-        .collection('users')
-        .doc(curUserUid)
-        .get()
-        .then((value) {
-      _profPicUrl = value.get('profPicUrl');
-    }).then((value) async {
-      if (_profPicUrl != "") {
-        final appStorage = await getApplicationDocumentsDirectory();
-        final file = File('${appStorage.path}/"profPic"');
-        final response = await Dio().get(_profPicUrl,
-            options: Options(
-                responseType: ResponseType.bytes,
-                followRedirects: false,
-                receiveTimeout: 0));
-        final raf = file.openSync(mode: FileMode.write);
-        raf.writeFromSync(response.data);
-        await raf.close();
-        _profPic = FileImage(file);
-      }
-      notifyListeners();
-    });
+  downloadProfPic() async {
+    final appStorage = await getApplicationDocumentsDirectory();
+    final file = File('${appStorage.path}/"profPic"');
+    if (!file.existsSync()) {
+      FirebaseFirestore.instance
+          .collection('users')
+          .doc(curUserUid)
+          .get()
+          .then((value) {
+        _profPicUrl = value.get('profPicUrl');
+      }).then((value) async {
+        if (_profPicUrl != "") {
+          final response = await Dio().get(_profPicUrl,
+              options: Options(
+                  responseType: ResponseType.bytes,
+                  followRedirects: false,
+                  receiveTimeout: 0));
+          final raf = file.openSync(mode: FileMode.write);
+          raf.writeFromSync(response.data);
+          await raf.close();
+          _profPic = FileImage(file);
+        }
+        notifyListeners();
+      });
+    }
   }
 
   imagePick(int source) async {
-    var imageSource;
+    ImageSource imageSource;
     if (source == 0) {
       imageSource = ImageSource.gallery;
     } else {
@@ -62,30 +66,72 @@ class ProfPicProvider extends ChangeNotifier {
     await ImageCropper.platform.cropImage(
         sourcePath: file.path,
         aspectRatio: const CropAspectRatio(ratioX: 1, ratioY: 1),
-        aspectRatioPresets: [CropAspectRatioPreset.original]).then((value) {
+        aspectRatioPresets: [
+          CropAspectRatioPreset.original
+        ]).then((value) async {
       if (value != null) {
+        _isUploading = true;
         _profPic = FileImage(File(value.path));
-        FirebaseStorageUtils().uploadImage(File(value.path));
+        uploadImage(File(value.path));
         notifyListeners();
       }
     });
   }
 
+  loadProfPic() async {
+    final appStorage = await getApplicationDocumentsDirectory();
+    final file = File('${appStorage.path}/"profPic"');
+    if (file.existsSync()) {
+      _profPic = FileImage(file);
+      notifyListeners();
+    }
+  }
+
   deleteProfPic() async {
-    final desertRef = FirebaseStorage.instance.ref('profPics/$curUserUid');
-    await desertRef.delete().then((value) {
-      FirebaseFirestore.instance
-          .collection('users')
-          .doc(curUserUid)
-          .get()
-          .then((value) async {
-        value.data()!.update('profPicUrl', (value) => "");
-        _profPicUrl = "";
-        _profPic = const AssetImage("res/images/defaultProf.png");
-        final appStorage = await getApplicationDocumentsDirectory();
-        final file = File('${appStorage.path}/"profPic"');
-        file.delete();
-        notifyListeners();
+    final appStorage = await getApplicationDocumentsDirectory();
+    final file = File('${appStorage.path}/"profPic"');
+    if (file.existsSync()) {
+      final desertRef = FirebaseStorage.instance.ref('profPics/$curUserUid');
+      await desertRef.delete().then((value) {
+        FirebaseFirestore.instance
+            .collection('users')
+            .doc(curUserUid)
+            .get()
+            .then((value) async {
+          value.reference.update({'profPicUrl': ""});
+          _profPicUrl = "";
+          _profPic = const AssetImage("res/images/defaultProf.png");
+          file.delete();
+          notifyListeners();
+        });
+      });
+    }
+  }
+
+  uploadImage(dynamic image) async {
+    if (image == null) return;
+    final destination = 'profPics/$curUserUid';
+    final ref = FirebaseStorage.instance.ref(destination);
+    ref.putFile(image).then((p0) async {
+      p0.ref.getDownloadURL().then((value) async {
+        await FirebaseFirestore.instance
+            .collection('users')
+            .doc(curUserUid)
+            .update({'profPicUrl': value}).then((value1) async {
+          _profPicUrl = value;
+          _isUploading = false;
+          final appStorage = await getApplicationDocumentsDirectory();
+          final file = File('${appStorage.path}/"profPic"');
+          final response = await Dio().get(_profPicUrl,
+              options: Options(
+                  responseType: ResponseType.bytes,
+                  followRedirects: false,
+                  receiveTimeout: 0));
+          final raf = file.openSync(mode: FileMode.write);
+          raf.writeFromSync(response.data);
+          await raf.close();
+          notifyListeners();
+        });
       });
     });
   }
